refactor(channels): flatten channel turn dispatch naming
