+++ src/channels/turn/lifecycle.ts
@@ -658,5 +658,3 @@ export async function dispatchRoutedChannelTurn(
     "routed-delivery",
   );
 }
-
-export { runPreparedInboundReply } from "./execution.js";
