+++ src/channels/turn/message-turn-guardrails.test.ts
@@ -2,9 +2,10 @@
 import fs from "node:fs";
 import path from "node:path";
 import { fileURLToPath } from "node:url";
-import { describe, expect, it } from "vitest";
+import { describe, expect, it, vi } from "vitest";
 import { expectNoReaddirSyncDuring } from "../../test-utils/fs-scan-assertions.js";
 import { listGitTrackedFiles } from "../../test-utils/repo-files.js";
+import { runChannelTurn } from "./run-channel-turn.js";
 
 const repoRoot = path.resolve(path.dirname(fileURLToPath(import.meta.url)), "../../..");
 
@@ -124,6 +125,22 @@ function collectReplyHistoryBindings(source: string): Set<string> {
 }
 
 describe("message turn migration guardrails", () => {
+  it("drops when ingest returns null", async () => {
+    const result = await runChannelTurn({
+      channel: "test",
+      raw: {},
+      adapter: {
+        ingest: () => null,
+        resolveTurn: vi.fn(),
+      },
+    });
+
+    expect(result).toEqual({
+      admission: { kind: "drop", reason: "ingest-null" },
+      dispatched: false,
+    });
+  });
+
   it("lists plugin TypeScript files from git without walking extension roots", () => {
     expectNoReaddirSyncDuring(() => {
       const files = listTsFiles("extensions");
