+++ src/plugin-sdk/channel-outbound.ts
@@ -6,7 +6,7 @@ import type {
 } from "../channels/message/runtime.js";
 import { createLazyRuntimeModule } from "../shared/lazy-runtime.js";
 
-type ChannelInboundKernelModule = typeof import("../channels/turn/kernel.js");
+type ChannelDurableDeliveryModule = typeof import("../channels/turn/durable-delivery.js");
 // Share one lazy import across SDK helper calls so plugin barrels do not eagerly pull
 // message runtime internals into registration/discovery-only paths.
 const loadChannelMessageRuntimeModule = createLazyRuntimeModule(
@@ -172,9 +172,9 @@ export type {
 } from "../channels/message/index.js";
 
 /** Lazily forwards inbound reply delivery through the channel turn kernel. */
-export const deliverInboundReplyWithMessageSendContext: ChannelInboundKernelModule["deliverInboundReplyWithMessageSendContext"] =
+export const deliverInboundReplyWithMessageSendContext: ChannelDurableDeliveryModule["deliverInboundReplyWithMessageSendContext"] =
   async (...args) => {
-    const mod = await import("../channels/turn/kernel.js");
+    const mod = await import("../channels/turn/durable-delivery.js");
     return await mod.deliverInboundReplyWithMessageSendContext(...args);
   };
