+++ src/plugins/runtime/types-channel.ts
@@ -184,12 +184,12 @@ export type PluginRuntimeChannel = {
   };
   inbound: {
     buildContext: typeof import("../../channels/inbound-event/context.js").buildChannelInboundEventContext;
-    run: typeof import("../../channels/turn/kernel.js").runChannelInboundEvent;
+    run: typeof import("../../channels/turn/run-channel-turn.js").runChannelTurn;
     /** @deprecated Prefer `run` for raw inbound events or `dispatchReply` for assembled contexts. */
-    runPreparedReply: typeof import("../../channels/turn/kernel.js").runPreparedInboundReply;
-    dispatch: typeof import("../../channels/turn/kernel.js").dispatchChannelInboundTurn;
+    runPreparedReply: typeof import("../../channels/turn/execution.js").runPreparedChannelTurn;
+    dispatch: typeof import("../../channels/turn/lifecycle.js").dispatchRoutedChannelTurn;
     /** Compatibility escape hatch; prefer `dispatch`, which keeps session wiring in core. */
-    dispatchReply: typeof import("../../channels/turn/kernel.js").dispatchChannelInboundReply;
+    dispatchReply: typeof import("../../channels/turn/lifecycle.js").dispatchAssembledChannelTurn;
   };
   threadBindings: {
     setIdleTimeoutBySessionKey: (params: {
