+++ test/scripts/check-deprecated-api-usage.test.ts
@@ -74,7 +74,6 @@ describe("scripts/check-deprecated-api-usage", () => {
       "src/plugin-sdk/channel-message",
       "src/plugin-sdk/channel-reply-pipeline",
       "src/plugin-sdk/inbound-reply-dispatch",
-      "src/channels/message/inbound-reply-dispatch",
     ]) {
       expect(modulePaths.has(facade), facade).toBe(true);
     }
@@ -94,7 +93,7 @@ describe("scripts/check-deprecated-api-usage", () => {
     const result = runFacadeImportRule({
       "src/channels/probe.ts": [
         'import { createChannelReplyPipeline } from "openclaw/plugin-sdk/channel-reply-pipeline";',
-        'export { runChannelInboundEvent } from "./message/inbound-reply-dispatch.js";',
+        'export { runChannelInboundEvent } from "../plugin-sdk/inbound-reply-dispatch.js";',
         'const facade = await import ("../plugin-sdk/channel-message.js", { with: {} });',
       ].join("\n"),
     });
@@ -103,14 +102,16 @@ describe("scripts/check-deprecated-api-usage", () => {
     expect(result.stderr).toContain(
       "src/channels/probe.ts:1: openclaw/plugin-sdk/channel-reply-pipeline",
     );
-    expect(result.stderr).toContain("src/channels/probe.ts:2: ./message/inbound-reply-dispatch.js");
+    expect(result.stderr).toContain(
+      "src/channels/probe.ts:2: ../plugin-sdk/inbound-reply-dispatch.js",
+    );
     expect(result.stderr).toContain("src/channels/probe.ts:3: ../plugin-sdk/channel-message.js");
   });
 
   it("allows canonical compat re-exports and test files", () => {
     const result = runFacadeImportRule({
-      "src/plugin-sdk/channel-inbound.ts":
-        'export { runChannelInboundEvent } from "../channels/message/inbound-reply-dispatch.js";',
+      "src/plugin-sdk/inbound-reply-dispatch.ts":
+        'export { runChannelInboundEvent } from "./channel-inbound.js";',
       "src/plugin-sdk/channel-message.test.ts":
         'const mod = await import("openclaw/plugin-sdk/channel-message");',
     });
